+++ music/musicPlayer.py
@@ -56,7 +56,7 @@ def textToSeconds(text):
     return seconds
 
 class Player():
-    def __init__(self, vc : discord.VoiceClient, queue : Queue) -> None:
+    def __init__(self, vc : discord.VoiceClient, queue : Queue, urlsync: list[dict]) -> None:
         mPrint('DEBUG', 'called Player __init__')
         self.queue : Queue = queue
         self.currentTrack : Track = None
@@ -65,6 +65,8 @@ class Player():
         self.isConnected = True
         self.timeout = config.no_music_timeout
         self.isWaiting = False
+
+        self.urlsync = urlsync
 
         #flags needed to communicate with EmbedHandler
         self.wasReported = False
@@ -147,7 +149,8 @@ class Player():
         if self.isPaused: self.resume()
 
         try:
-            song_url = self.currentTrack.getVideoUrl()
+            mPrint('TEST', f"{self.urlsync=}")
+            song_url = self.currentTrack.getVideoUrl(urlsync=self.urlsync)
             mPrint('DEBUG', f'URL: {song_url}')
             if song_url == None:
                 # careful with recursion
@@ -224,6 +227,9 @@ class Player():
         self.voiceClient.resume()
         self.isPaused = False
         self.pauseStart = time.time()
+
+    def remove(self, index = 0):
+        self.queue.removeAtIndex(index)
 
     def clear(self): #kind of does not make sense this just seems like stop() but spicy
         self.queue.clear()
@@ -509,7 +515,7 @@ class MessageHandler():
 
         if self.currentTrack == None: return
 
-        if self.currentTrack.getSource() == 'spotify':
+        if self.currentTrack.source == 'spotify':
             #do not update the buttons if there is nothing to change
             if self.currentTrack.title in self.spotifyBtn.label: return
 
@@ -518,7 +524,7 @@ class MessageHandler():
             if len(btnlabel) > 80:
                 btnlabel = "Ascoltala su spotify"
             self.spotifyBtn.label = btnlabel
-            self.spotifyBtn.url = self.currentTrack.getOriginalURL()
+            self.spotifyBtn.url = self.currentTrack.spotifyURL
             self.spotifyBtn.disabled = False
             self.view.add_item(self.spotifyBtn)
         else: